[cleanup] Move `renderJoin` out
- Move all join-related code from `CFunctionGeneration` out to the
`com.nec.spark.agile.join` package

- Move `GroupingCodeGenerator` over to the `com.nec.spark.agile.groupby` package
