+++ src/main/scala/com/nec/spark/agile/groupby/GroupByOutline.scala
@@ -22,7 +22,6 @@ package com.nec.spark.agile.groupby
 import com.nec.spark.agile.core.CodeLines
 import com.nec.spark.agile.CFunctionGeneration._
 import com.nec.spark.agile.groupby.GroupByOutline.{GroupingKey, StagedAggregation, StagedProjection}
-import com.nec.spark.agile.{GroupingCodeGenerator}
 
 /**
  * General class to describe a group-by to create the function outline
