+++ src/main/scala/com/nec/spark/planning/VeFunction.scala
@@ -22,7 +22,7 @@ final case class VeFunction(
   functionName: String,
   namedResults: List[CVector]
 ) {
-  def results: List[VeType] = namedResults.map(_.veType)
+  def results: Seq[VeType] = namedResults.map(_.veType)
   def isCompiled: Boolean = veFunctionStatus match {
     case VeFunctionStatus.SourceCode(_) => false
     case VeFunctionStatus.Compiled(_)   => true
