+++ src/main/scala/com/nec/ve/VeKernelCompiler.scala
@@ -177,7 +177,6 @@ final case class VeKernelCompiler(
           oFile.toString
         )
       logger.info(s"Compilation command = ${command}")
-      println(s"Compilation command = ${command}")
 
       ProcessRunner.runHopeOk(
         Process(command = command, cwd = buildDir.toFile),
@@ -197,7 +196,6 @@ final case class VeKernelCompiler(
         ) ++ linkSos.toList.map(_.name).map(sourcesDir.resolve(_)).map(_.toString)
       }
       logger.info(s"Compilation command 2 = ${command2}")
-      println(s"Compilation command 2 = ${command2}")
 
       ProcessRunner.runHopeOk(
         Process(command = command2, cwd = buildDir.toFile),
