+++ src/test/scala/com/nec/cmake/eval/OldUnifiedGroupByFunctionGeneration.scala
@@ -20,25 +20,11 @@
 package com.nec.cmake.eval
 
 import com.nec.spark.agile.core.CodeLines
-import com.nec.spark.agile.CFunctionGeneration.GroupByExpression.{
-  GroupByAggregation,
-  GroupByProjection
-}
+import com.nec.spark.agile.CFunctionGeneration.GroupByExpression.{GroupByAggregation, GroupByProjection}
 import com.nec.spark.agile.CFunctionGeneration._
-import com.nec.spark.agile.GroupingCodeGenerator
 import com.nec.spark.agile.StringProducer.CopyStringProducer
-import com.nec.spark.agile.groupby.GroupByOutline.{
-  GroupingKey,
-  StagedAggregation,
-  StagedAggregationAttribute,
-  StagedProjection,
-  StringReference
-}
-import com.nec.spark.agile.groupby.{
-  GroupByOutline,
-  GroupByPartialGenerator,
-  GroupByPartialToFinalGenerator
-}
+import com.nec.spark.agile.groupby.GroupByOutline.{GroupingKey, StagedAggregation, StagedAggregationAttribute, StagedProjection, StringReference}
+import com.nec.spark.agile.groupby.{GroupByOutline, GroupByPartialGenerator, GroupByPartialToFinalGenerator, GroupingCodeGenerator}
 import com.nec.spark.planning.VERewriteStrategy.SequenceList
 
 final case class OldUnifiedGroupByFunctionGeneration(
